Move platform-specific code around
- Priority-setting code: main -> platform
- Low-level code: sys -> platform/sys
