+++ platform/priority_linux.go
@@ -1,15 +1,15 @@
 // +build linux
 
-package main
+package platform
 
 import (
 	"github.com/spyre-project/spyre/log"
-	"github.com/spyre-project/spyre/sys"
+	"github.com/spyre-project/spyre/platform/sys"
 
 	"syscall"
 )
 
-func setLowPriority() {
+func SetLowPriority() {
 	pid := syscall.Getpid()
 	if err := syscall.Setpriority(syscall.PRIO_PROCESS, pid, 10); err != nil {
 		log.Errorf("Failed to set CPU priority: %v", err)
