+++ platform/priority_unsupported.go
@@ -1,13 +1,14 @@
 // +build !darwin,!dragonfly,!freebsd,!linux,!netbsd,!openbsd,!solaris,!windows
 
-package main
+package platform
 
 import (
 	"github.com/spyre-project/spyre/log"
+
 	"runtime"
 )
 
-func setLowPriority() {
+func SetLowPriority() {
 	log.Error("priority setting is not supported on " + runtime.GOOS)
 	return
 }
