+++ platform/sys/syscall_linux.go
